fixed_package_names
