+++ app/internal/handlers/app/interface/interface.go
@@ -1,4 +1,4 @@
-package interfacehandler
+package appinterfacehandler
 
 import (
 	"app/internal/public/views/interface"
